Feature/add strategies

See merge request elabank/quicksilver/smartwallet!2

+++ contracts/GlobalConfig.sol
@@ -8,17 +8,24 @@ import "@openzeppelin/contracts/access/Ownable.sol";
  * @dev Config contract shared by all smart wallet contracts
  */
 contract GlobalConfig is Ownable {
-    uint public feeRatio;
+    uint256 public feeRatio;
     // each prod (such as ptd) has one strategyFactory to create strategy
     mapping(string => address) public tokenStrategyFactory;
 
     event NewStrategyFactory(string, address);
 
-    function getStrategyFactory(string calldata productName) external view returns (address) {
+    function getStrategyFactory(string calldata productName)
+        external
+        view
+        returns (address)
+    {
         return tokenStrategyFactory[productName];
     }
 
-    function setStrategyFactory(string calldata productName, address strategyFactory) external onlyOwner {
+    function setStrategyFactory(
+        string calldata productName,
+        address strategyFactory
+    ) external onlyOwner {
         require(strategyFactory != address(0), "GlobalConfig: zero address");
 
         tokenStrategyFactory[productName] = strategyFactory;
@@ -26,7 +33,7 @@ contract GlobalConfig is Ownable {
         emit NewStrategyFactory(productName, strategyFactory);
     }
 
-    function setFeeRatio(uint _feeRatio) external onlyOwner {
+    function setFeeRatio(uint256 _feeRatio) external onlyOwner {
         feeRatio = _feeRatio;
     }
 }
